Redefined how the `run` configuration option executes
As on-the-fly analysis may become very slow for certain static
analysis tools; the `run` configuration option has been
changed to permit on-the-fly analysis when specific analyzers
support such operation. When an analyzer cannot support on-the-
fly analysis, it is then performed at the time of save.

The option for at-time-of-save analysis has not changed.