Addon support for cppcheck
You can now pass cppcheck addons to this extension which are used for producing even more lint errors. Some slight modifications how unique diagnostic messages are handled when the error comes from the misra addon. This is because the addon produces identical messages for each vialation, but does have unieque codes. The misra addon does not provide the actual error but only the error code and you need to own the misra guidelines to actually know the rule vialation.