Compile and Debug with recent versions of VS Code and Node (#69)
* Latest Version of Typescript as dev Dependency

Upgraded to the latest version of typescript as a dev dependency. The reason for this is that with the older version, the project could not compile correctly (node 12).

* Debug environment compatibility with Node 12 and VS Code 1.41

Npm packages have been updated and some modifications in the code in accordance with non backwards-compatible npm packages. This fixes a problem mentioned in issue #59, which is that you get the error message "Starting client failed with error code 9.

* Changed to node version 12 in CI

Signed-off-by: TSonono <tofiksonono@msn.com>