Multiple violations of the same type not displayed
The linter was not able to display violations of the same type that appe
ared on different lines. This commit contains a fix for this aswell as a
n unit test for this scenario. The problem however, is that the unit tes
t will not detect this error since in the test scope, the output has not
passed through the uniqBy filtering.

The fix to this problem was to "re-add" the line check  to the lambda in
uniqBy. However, this time it is done with e.range.start.line instead of
e.line. This field exists and carries the correct value of the line.

Signed-off-by: TSonono <tofiksonono@msn.com>