cppcheck: add 'extraArgs' argument
Let users pass extra command line arguments to cppcheck via the
'extraArgs' setting in the client.

Works the same as 'extraArgs' for clang.

This is helpful so users can specify other options that the
settings does not cover, such as "--project", in order to
point cppcheck to a compilation database and "-j", in order
to invoke cppcheck with multiple jobs.

Signed-off-by: Daniel W. S. Almeida <dwlsalmeida@gmail.com>