Allow diagnostic code lookup on the web
In some cases, like for misra rule checking with CppCheck, no useful
description is available from the tool. This change introduces the
ability to search the diagnostic code on the web for more information.
The query base Uri and the set of words to match to activate the
behavior are configurable through settings. For security reasons, the
query base Uri is only configurable from trusted workspaces.
The feature also only activates when the language server doesn't
already provide a link in the generated diagnostic.